Principal component analysis for bigmemory::big.matrix inputs

Perform principal component analysis (PCA) directly on a bigmemory::big.matrix without copying the data into R memory. The exported helpers mirror the structure of base R's prcomp() while avoiding the need to materialise large matrices.

Usage

resolve_big_pointer(x, arg, allow_null = FALSE)

Arguments

xpMat

Either a bigmemory::big.matrix or an external pointer such as mat@address that references the source big.matrix.

center

Logical; should column means be subtracted before PCA?

scale

Logical; should variables be scaled to unit variance? Scaling requires center = TRUE.

ncomp

Number of components to retain. Use a non-positive value to keep all components returned by the decomposition.

block_size

Number of rows to process per block when streaming data through BLAS kernels. Larger values improve throughput at the cost of additional memory.

Value

For pca_bigmatrix(), a bigpca object mirroring a prcomp result with elements sdev, rotation, optional center and scale vectors, column_sd, eigenvalues, explained_variance, cumulative_variance, and the sample covariance matrix. The object participates in S3 generics such as summary() and plot().

See also

Examples

set.seed(123)
mat <- bigmemory::as.big.matrix(matrix(rnorm(40), nrow = 10))
pca <- pca_bigmatrix(mat, center = TRUE, scale = TRUE, ncomp = 3)
scores <- pca_scores_bigmatrix(mat, pca$rotation, pca$center, pca$scale, ncomp = 3)
loadings <- pca_variable_loadings(pca$rotation, pca$sdev)
correlations <- pca_variable_correlations(pca$rotation, pca$sdev, pca$column_sd)
contributions <- pca_variable_contributions(loadings)
list(scores = scores, loadings = loadings, correlations = correlations,
     contributions = contributions)
#> $scores
#>              [,1]        [,2]        [,3]
#>  [1,]  0.54325339 -0.91493729 -0.41315014
#>  [2,] -0.77428157 -0.84150780  0.40023111
#>  [3,]  1.77362948  0.79996096  0.05039649
#>  [4,]  0.61761772  0.59591822 -0.57031347
#>  [5,]  0.23218256  0.95503144 -0.72268699
#>  [6,]  2.66813521 -0.45268284  0.33775240
#>  [7,] -0.08448606  0.81203899  1.15087152
#>  [8,] -2.43891736  0.54722825 -0.96050148
#>  [9,] -0.40827305 -1.59689839 -0.36789931
#> [10,] -2.12886032  0.09584844  1.09529988
#> 
#> $loadings
#>            [,1]       [,2]       [,3]
#> [1,]  0.8567704  0.2711956  0.3439932
#> [2,]  0.7678169 -0.5050498  0.3324763
#> [3,] -0.7573969  0.3708264  0.5203357
#> [4,]  0.7581891  0.5754452 -0.2056261
#> 
#> $correlations
#>            [,1]       [,2]       [,3]
#> [1,]  0.8982856  0.2843364  0.3606615
#> [2,]  0.7396557 -0.4865261  0.3202821
#> [3,] -0.8136972  0.3983915  0.5590143
#> [4,]  1.4378639  1.0913001 -0.3899587
#> 
#> $contributions
#>           [,1]       [,2]       [,3]
#> [1,] 0.2969361 0.09224839 0.21836250
#> [2,] 0.2384786 0.31993524 0.20398568
#> [3,] 0.2320498 0.17247851 0.49962661
#> [4,] 0.2325354 0.41533786 0.07802521
#>